refactore computer to composition
