+++ _05-OOP-2_polymorphism/src/Composition/Main.java
@@ -1,4 +1,4 @@
-package Computer;
+package Composition;
 public class Main {
     public static void main(String[] args) {
         ComputerCase theCase = new ComputerCase("2208", "Dell", "240");
