+++ _05-OOP-2_polymorphism/src/Composition/Product.java
@@ -1,4 +1,4 @@
-package Computer;
+package Composition;
 
 public class Product {
     private String model;
