+++ _05-OOP-2_polymorphism/src/Composition/challenge/Main.java
@@ -1,4 +1,4 @@
-package CompositionChallenge;
+package Composition.challenge;
 
 public class Main {
     public static void main(String[] args) {
