+++ _05-OOP-2_polymorphism/src/Composition/challenge/SmartKitchen.java
@@ -1,4 +1,4 @@
-package CompositionChallenge;
+package Composition.challenge;
 
 public class SmartKitchen {
     private CoffeeMaker brewMaster;
